+++ CHANGELOG.md
@@ -5,9 +5,9 @@ All notable changes to the `cynthion` Python package will be documented in this 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!--
 ## [Unreleased]
--->
+### Added
+* New `cynthion reset` command to reset the FPGA and reconfigure it from flash. (#216)
 
 ## [0.2.5] - 2026-05-22
 ### Added